feat: support process locking
Adds a program lock for long running process such as installations, uninstalls and updates as these could have issues if they overlap on eachothers execiton. The lock does not actually need to be used on instant returns as they wont have any issue with other processes. The GUI should trigger a lock though, however maybe in the future this should use a different lock file to only allow 1 gui at a time, and use the core lock file for everything else

Signed-off-by: BitsOfAByte <19539165+BitsOfAByte@users.noreply.github.com>